requirements.txt
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Flask==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
requests==2.32.4
urllib3==2.5.0
Werkzeug==3.1.3
python-dotenv==1.0.0
slack_bolt==1.23.0
slack_sdk==3.35.0
